Added qualimap module
